chore: move test_extcodesize_bytecode_sizes to test_account_query
